requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
gunicorn==21.2.0
Flask-Login==0.6.3
Flask-Session==0.5.0
requests==2.31.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7
PyPDF2==3.0.1
python-docx==0.8.11
openai==1.3.0 